+++ src/integrationTest/java/uk/gov/hmcts/reform/bulkscan/orchestrator/controllers/AttachExceptionRecordToExistingCaseTest.kt
@@ -4,6 +4,10 @@ import io.restassured.RestAssured
 import io.restassured.RestAssured.given
 import io.restassured.builder.RequestSpecBuilder
 import io.restassured.http.ContentType.JSON
+import io.restassured.response.Response
+import io.restassured.response.ValidatableResponse
+import io.restassured.response.ValidatableResponseOptions
+import io.restassured.specification.RequestSpecification
 import org.hamcrest.Matchers.contains
 import org.hamcrest.Matchers.equalTo
 import org.junit.jupiter.api.BeforeEach
@@ -14,8 +18,15 @@ import org.springframework.test.context.junit.jupiter.SpringExtension
 import u.gov.hmcts.reform.bulkscan.orchestrator.controllers.config.PortWaiter.waitFor
 import uk.gov.hmcts.reform.bulkscan.orchestrator.controllers.config.IntegrationTest
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest
+import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest.CallbackRequestBuilder
 import uk.gov.hmcts.reform.ccd.client.model.CallbackTypes
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails
+
+typealias ResponseValidation = ValidatableResponseOptions<ValidatableResponse, Response>
+
+fun RequestSpecification.postToCallback(type: String = "attach_case") = post("/callback/{type}", type)
+fun RequestSpecification.setBody(builder: CallbackRequestBuilder) = body(builder.build())
+fun ResponseValidation.shouldContainError(error: String) = body("errors", contains(error))
 
 @ExtendWith(SpringExtension::class)
 @IntegrationTest
@@ -38,8 +49,8 @@ class AttachExceptionRecordToExistingCaseTest {
     @Test
     fun `should successfully callback with correct information`() {
         given()
-            .body(request.build())
-            .post("/callback/{type}", "attach_case")
+            .setBody(request)
+            .postToCallback()
             .then()
             .statusCode(200)
             .body("errors.size()", equalTo(0))
@@ -48,40 +59,40 @@ class AttachExceptionRecordToExistingCaseTest {
     @Test
     fun `should fail with the correct error when no case data is supplied`() {
         given()
-            .body(request.caseDetails(null).build())
-            .post("/callback/{type}", "attach_case")
+            .setBody(request.caseDetails(null))
+            .postToCallback()
             .then()
             .statusCode(200)
-            .body("errors", contains("Internal Error: no Case details supplied"))
+            .shouldContainError("Internal Error: no Case details supplied")
     }
 
     @Test
     fun `should fail if invalid eventId set`() {
         given()
-            .body(request.eventId("invalid").build())
-            .post("/callback/{type}", "attach_case")
+            .setBody(request.eventId("invalid"))
+            .postToCallback()
             .then()
             .statusCode(200)
-            .body("errors", contains("Internal Error: event-id: invalid invalid"))
+            .shouldContainError("Internal Error: event-id: invalid invalid")
     }
 
     @Test
     fun `should fail if no eventId set`() {
         given()
-            .body(request.eventId(null).build())
-            .post("/callback/{type}", "attach_case")
+            .setBody(request.eventId(null))
+            .postToCallback()
             .then()
             .statusCode(200)
-            .body("errors", contains("Internal Error: event-id: null invalid"))
+            .shouldContainError("Internal Error: event-id: null invalid")
     }
 
     @Test
     fun `should create error if type in incorrect`() {
         given()
-            .body(request.build())
-            .post("/callback/{type}", "someType")
+            .setBody(request)
+            .postToCallback("someType")
             .then()
             .statusCode(200)
-            .body("errors", contains("Internal Error: invalid type supplied: someType"))
+            .shouldContainError("Internal Error: invalid type supplied: someType")
     }
 }
