+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/CallbackProcessorImpl.java
@@ -2,27 +2,18 @@ package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd;
 
 import io.vavr.Value;
 import io.vavr.control.Validation;
-import org.jetbrains.annotations.NotNull;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 
 import java.util.List;
 
-import static io.vavr.control.Validation.invalid;
-import static io.vavr.control.Validation.valid;
-import static java.lang.String.format;
 import static java.util.Collections.emptyList;
-import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackProcessorImpl.CallbackValidations.hasCaseDetails;
-import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackProcessorImpl.CallbackValidations.internalError;
-import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackProcessorImpl.CallbackValidations.isAboutToSubmit;
-import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackProcessorImpl.CallbackValidations.isAttachEvent;
-import static uk.gov.hmcts.reform.ccd.client.model.CallbackTypes.ABOUT_TO_SUBMIT;
+import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackValidations.hasCaseDetails;
+import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackValidations.isAboutToSubmit;
+import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackValidations.isAttachEvent;
 
 @Service
 public class CallbackProcessorImpl implements CallbackProcessor {
-    private static Logger log = LoggerFactory.getLogger(CallbackProcessor.class);
 
     @Override
     public List<String> process(String eventType, String eventId, CaseDetails caseDetails) {
@@ -36,36 +27,8 @@ public class CallbackProcessorImpl implements CallbackProcessor {
             .getOrElseGet(Value::toJavaList);
     }
 
-    List<String> attach(CaseDetails caseDetails) {
+    private List<String> attach(CaseDetails caseDetails) {
         return emptyList();
-    }
-
-    interface CallbackValidations {
-
-        static Validation<String, CaseDetails> hasCaseDetails(CaseDetails caseDetails) {
-            return caseDetails != null
-                ? valid(caseDetails)
-                : internalError("no Case details supplied", null);
-        }
-
-        static Validation<String, String> isAboutToSubmit(String eventId) {
-            return ABOUT_TO_SUBMIT.equals(eventId)
-                ? valid(eventId)
-                : internalError("event-id: %s invalid", eventId);
-        }
-
-        @NotNull
-        static <T> Validation<String, T> internalError(String error, T arg1) {
-            log.error("{}:{}", error, arg1);
-            return invalid(format("Internal Error: " + error, arg1));
-        }
-
-
-        static Validation<String, String> isAttachEvent(String type) {
-            return "attach_case".equals(type)
-                ? valid(type)
-                : internalError("invalid type supplied: %s", type);
-        }
     }
 
 }
